Use simple elaboration strategy for ap, transport, etc.
This solves the following problem mentioned in https://github.com/gebner/hott3/pull/1

> `transport (λa : A, P a b) p x` [does not elaborate]

This is an unfortunate consequence of a second-order unification heuristic in the type context, I think.  For this expression, the elaborator has the expected type `P a b` and it tries to unify it with the result type of `transport`---which is `?m_1 ?m_2`.  Now we have the unification problem `?m_1 ?m_2 =?= P a b`, and we heuristically assign `?m_1 := P a`---which is of course completely wrong here.

Similar problems arise with `ap`, `pathover`, etc...